rename and move binding module for Activity
